Refactored ts files names from lowerCamelCase to CamelCase
