+++ src/main/resources/static/ts/UiManager.ts
@@ -1,16 +1,15 @@
-import {Header} from "./ui/header";
-import {MessagesPage} from "./ui/messagesPage";
-import {PageOfUser} from "./ui/pageOfUser";
-import {SearchPage} from "./ui/searchPage";
-import {IndexPage} from "./ui/indexPage";
-import {RegistrationPage} from "./ui/registrationPage";
-import {AbstractPage} from "./utils/abstractPage";
+import {Header} from "./ui/Header";
+import {MessagesPage} from "./ui/MessagesPage";
+import {PageOfUser} from "./ui/PageOfUser";
+import {SearchPage} from "./ui/SearchPage";
+import {IndexPage} from "./ui/IndexPage";
+import {RegistrationPage} from "./ui/RegistrationPage";
+import {AbstractPage} from "./utils/AbstractPage";
 import * as $ from "jquery";
 import jqXHR = JQuery.jqXHR;
-import {SpendingsPage} from "./ui/spendingsPage";
-import { NewPearsonPage } from "./ui/newPearsonPage";
+import {NewPearsonPage} from "./ui/NewPearsonPage";
 
-class UIManager {
+class UiManager {
 
     private header: Header = null;
     private page: AbstractPage;
@@ -36,22 +35,22 @@ class UIManager {
     }
 
     public getPage(pageObject): void {
-        if(pageObject.pageName !== "registration" && pageObject.pageName !== "login"){
-        $.ajax(
-            {
-                url: "/authentication",
-                type: "GET",
-            }
-        ).then((data) => {
-            pageObject.pageName = data === "authenticated" ? pageObject.pageName : "login";
-            this.loadPage(pageObject);
-        }).catch((e: jqXHR) => {
-            console.log(e);
-            if (e.status >= 400) {
-                pageObject.pageName = "login";
+        if (pageObject.pageName !== "registration" && pageObject.pageName !== "login") {
+            $.ajax(
+                {
+                    url: "/authentication",
+                    type: "GET",
+                }
+            ).then((data) => {
+                pageObject.pageName = data === "authenticated" ? pageObject.pageName : "login";
                 this.loadPage(pageObject);
-            }
-        });
+            }).catch((e: jqXHR) => {
+                console.log(e);
+                if (e.status >= 400) {
+                    pageObject.pageName = "login";
+                    this.loadPage(pageObject);
+                }
+            });
         } else {
             this.loadPage(pageObject);
         }
@@ -76,12 +75,6 @@ class UIManager {
                 this.renderHeader();
                 history.pushState("", "", `application.html?search`);
                 this.page = new SearchPage(this.myId, this.myUsername);
-                break;
-            }
-            case "spendings": {
-                this.renderHeader();
-                history.pushState("","", `application.html?spendings`);
-                this.page = new SpendingsPage(this.myId, this.myUsername);
                 break;
             }
             case "login": {
@@ -116,16 +109,16 @@ class UIManager {
         }
     }
 
-    private onWindowResize(){
-        if(document.location.search.indexOf("messages") > -1){
+    private onWindowResize() {
+        if (document.location.search.indexOf("messages") > -1) {
             (document.querySelector("div.inbox_msg") as HTMLDivElement).style.cssText = `width: 900px; height: ${$(window).height() - 60}px;`;
             (document.querySelector("div.msg_history") as HTMLDivElement).style.cssText = `height: ${$(window).height() - 210}px;`;
-        }else if(document.location.search.indexOf("registration") > -1) {
+        } else if (document.location.search.indexOf("registration") > -1) {
             (document.querySelector("div#reg_form_block") as HTMLDivElement).style.cssText = `height: ${$(window).height() - 60}px;`;
-        }else if(document.location.search.indexOf("pageOfUser") > -1) {
+        } else if (document.location.search.indexOf("pageOfUser") > -1) {
             $("#user_page").css({"height": `${$(window).height() - 60}px`});
         }
     }
 }
 
-export const uiManager = new UIManager();
+export const uiManager = new UiManager();
