+++ src/main/resources/static/ts/utils/AbstractPage.ts
@@ -3,7 +3,7 @@ export abstract class AbstractPage {
     protected myId: string;
     protected myUsername: string;
 
-    protected constructor(myId:string, myUsername: string) {
+    protected constructor(myId: string, myUsername: string) {
         this.myId = myId;
         this.myUsername = myUsername;
     }
