+++ src/main/resources/static/ts/utils/Logger.ts
