+++ src/main/resources/static/ts/utils/templates/IDialog.ts
