+++ src/main/resources/static/ts/utils/templates/IMention.ts
