+++ src/main/resources/static/ts/utils/templates/IMessage.ts
