+++ src/main/resources/static/ts/utils/templates/INewPearson.ts
@@ -2,7 +2,7 @@ export interface INewPearson {
     first_name: string;
     second_name: string;
     last_name: string;
-    
+
     phone_number: string;
     country: string;
     city: string;
