+++ src/main/resources/static/ts/utils/templates/IUser.ts
